Move all in package 'userDB' to 'db' except DbUser.java. Avoids checkstyle issues.
