+++ src/integ_test/java/games/strategy/engine/lobby/server/db/DbUserControllerIntegrationTest.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.lobby.server.userdb;
+package games.strategy.engine.lobby.server.db;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
