+++ src/main/java/games/strategy/engine/lobby/server/db/UserDao.java
@@ -1,4 +1,4 @@
-package games.strategy.engine.lobby.server.userdb;
+package games.strategy.engine.lobby.server.db;
 
 import games.strategy.engine.lobby.server.userDB.DBUser;
