+++ src/main/java/games/strategy/engine/lobby/server/login/LobbyLoginValidator.java
@@ -14,12 +14,12 @@ import com.google.common.base.Strings;
 
 import games.strategy.engine.framework.startup.ui.InGameLobbyWatcher;
 import games.strategy.engine.lobby.server.LobbyServer;
+import games.strategy.engine.lobby.server.db.BadWordController;
+import games.strategy.engine.lobby.server.db.BannedMacController;
+import games.strategy.engine.lobby.server.db.BannedUsernameController;
+import games.strategy.engine.lobby.server.db.DbUserController;
+import games.strategy.engine.lobby.server.db.HashedPassword;
 import games.strategy.engine.lobby.server.userDB.DBUser;
-import games.strategy.engine.lobby.server.userdb.BadWordController;
-import games.strategy.engine.lobby.server.userdb.BannedMacController;
-import games.strategy.engine.lobby.server.userdb.BannedUsernameController;
-import games.strategy.engine.lobby.server.userdb.DbUserController;
-import games.strategy.engine.lobby.server.userdb.HashedPassword;
 import games.strategy.net.ILoginValidator;
 import games.strategy.util.MD5Crypt;
 import games.strategy.util.Tuple;
